Reorganized editing files
