+++ client/src/components/edit/clubs/edit-committee.js
@@ -5,10 +5,9 @@ import { Committee } from '../../../functions/entries';
 import ListItem from '@material-ui/core/ListItem';
 import Box from '@material-ui/core/Box';
 import Card from '@material-ui/core/Card';
-import TextField from '@material-ui/core/TextField';
-import LinkInputList from '../util/link-input-list';
-import TrashCan from '../util/trash-can';
-import ControlledTextField from '../util/controlled-text-field';
+import LinkInputList from './link-input-list';
+import TrashCan from './trash-can';
+import ControlledTextField from '../shared/controlled-text-field';
 
 const useStyles = makeStyles((theme) => ({
     root: {
