+++ client/src/components/edit/clubs/trash-can.js
