+++ client/src/components/edit/edit-login.js
@@ -7,7 +7,7 @@ import { getIp, getLoggedIn, getUserInfo } from '../../functions/api';
 import Box from '@material-ui/core/Box';
 import Typography from '@material-ui/core/Typography';
 import { darkSwitchGrey } from '../../functions/util';
-import GoogleLoginButton from './util/google-login-button';
+import GoogleLoginButton from './shared/google-login-button';
 import { openConnectionErrorPopup } from '../../redux/actions';
 
 const useStyles = makeStyles((theme) => ({
