+++ client/src/components/edit/shared/upload-backdrop.js
